+++ lib/services/consumptionManagement/lib/models/index.d.ts
@@ -106,10 +106,16 @@ export interface Resource extends BaseResource {
  * @member {number} [meterDetails.totalIncludedQuantity] The total included
  * quantity associated with the offer.
  * @member {number} [meterDetails.pretaxStandardRate] The pretax listing price.
- * @member {object} [additionalProperties] The list of key/value pairs for the
- * additional properties, in the format 'key':'value' where key = the field
- * name, and value = the field value. By default this is not populated, unless
- * it's specified in $expand.
+ * @member {string} [subscriptionGuid] Subscription guid.
+ * @member {string} [subscriptionName] Subscription name.
+ * @member {string} [accountName] Account name.
+ * @member {string} [departmentName] Department name.
+ * @member {string} [product] Product name.
+ * @member {string} [consumedService] Consumed service name.
+ * @member {string} [costCenter] The cost center of this department if it is a
+ * department and a costcenter exists
+ * @member {string} [additionalProperties] Additional details of this usage
+ * item. By default this is not populated, unless it's specified in $expand.
  */
 export interface UsageDetail extends Resource {
   readonly billingPeriodId?: string;
@@ -126,7 +132,96 @@ export interface UsageDetail extends Resource {
   readonly isEstimated?: boolean;
   readonly meterId?: string;
   readonly meterDetails?: MeterDetails;
-  readonly additionalProperties?: { [propertyName: string]: string };
+  readonly subscriptionGuid?: string;
+  readonly subscriptionName?: string;
+  readonly accountName?: string;
+  readonly departmentName?: string;
+  readonly product?: string;
+  readonly consumedService?: string;
+  readonly costCenter?: string;
+  readonly additionalProperties?: string;
+}
+
+/**
+ * @class
+ * Initializes a new instance of the ReservationSummaries class.
+ * @constructor
+ * reservation summaries resource.
+ *
+ * @member {string} [reservationOrderId] The reservation order ID is the
+ * identifier for a reservation purchase. Each reservation order ID represents
+ * a single purchase transaction. A reservation order contains reservations.
+ * The reservation order specifies the VM size and region for the reservations.
+ * @member {string} [reservationId] The reservation ID is the identifier of a
+ * reservation within a reservation order. Each reservation is the grouping for
+ * applying the benefit scope and also specifies the number of instances to
+ * which the reservation benefit can be applied to.
+ * @member {string} [skuName] This is the ARM Sku name. It can be used to join
+ * with the servicetype field in additoinalinfo in usage records.
+ * @member {number} [reservedHours] This is the total hours reserved. E.g. if
+ * reservation for 1 instance was made on 1 PM, this will be 11 hours for that
+ * day and 24 hours from subsequent days
+ * @member {date} [usageDate] Data corresponding to the utilization record. If
+ * the grain of data is monthly, it will be first day of month.
+ * @member {number} [usedHours] Total used hours by the reservation
+ * @member {number} [minUtilizationPercentage] This is the minimum hourly
+ * utilization in the usage time (day or month). E.g. if usage record
+ * corresponds to 12/10/2017 and on that for hour 4 and 5, utilization was 10%,
+ * this field will return 10% for that day
+ * @member {number} [avgUtilizationPercentage] This is average utilization for
+ * the entire time range. (day or month depending on the grain)
+ * @member {number} [maxUtilizationPercentage] This is the maximum hourly
+ * utilization in the usage time (day or month). E.g. if usage record
+ * corresponds to 12/10/2017 and on that for hour 4 and 5, utilization was
+ * 100%, this field will return 100% for that day.
+ */
+export interface ReservationSummaries extends Resource {
+  readonly reservationOrderId?: string;
+  readonly reservationId?: string;
+  readonly skuName?: string;
+  readonly reservedHours?: number;
+  readonly usageDate?: Date;
+  readonly usedHours?: number;
+  readonly minUtilizationPercentage?: number;
+  readonly avgUtilizationPercentage?: number;
+  readonly maxUtilizationPercentage?: number;
+}
+
+/**
+ * @class
+ * Initializes a new instance of the ReservationDetails class.
+ * @constructor
+ * reservation details resource.
+ *
+ * @member {string} [reservationOrderId] The reservation order ID is the
+ * identifier for a reservation purchase. Each reservation order ID represents
+ * a single purchase transaction. A reservation order contains reservations.
+ * The reservation order specifies the VM size and region for the reservations.
+ * @member {string} [reservationId] The reservation ID is the identifier of a
+ * reservation within a reservation order. Each reservation is the grouping for
+ * applying the benefit scope and also specifies the number of instances to
+ * which the reservation benefit can be applied to.
+ * @member {string} [skuName] This is the ARM Sku name. It can be used to join
+ * with the servicetype field in additoinalinfo in usage records.
+ * @member {number} [reservedHours] This is the total hours reserved for the
+ * day. E.g. if reservation for 1 instance was made on 1 PM, this will be 11
+ * hours for that day and 24 hours from subsequent days.
+ * @member {date} [usageDate] The date on which consumption occurred.
+ * @member {number} [usedHours] This is the total hours used by the instance.
+ * @member {string} [instanceId] This identifier is the name of the resource or
+ * the fully qualified Resource ID.
+ * @member {number} [totalReservedQuantity] This is the total count of
+ * instances that are reserved for the reservationid.
+ */
+export interface ReservationDetails extends Resource {
+  readonly reservationOrderId?: string;
+  readonly reservationId?: string;
+  readonly skuName?: string;
+  readonly reservedHours?: number;
+  readonly usageDate?: Date;
+  readonly usedHours?: number;
+  readonly instanceId?: string;
+  readonly totalReservedQuantity?: number;
 }
 
 /**
@@ -138,12 +233,10 @@ export interface UsageDetail extends Resource {
  * @member {string} [code] Error code.
  * @member {string} [message] Error message indicating why the operation
  * failed.
- * @member {string} [target] The target of the particular error.
  */
 export interface ErrorDetails {
   readonly code?: string;
   readonly message?: string;
-  readonly target?: string;
 }
 
 /**
@@ -157,7 +250,6 @@ export interface ErrorDetails {
  * @member {string} [error.code] Error code.
  * @member {string} [error.message] Error message indicating why the operation
  * failed.
- * @member {string} [error.target] The target of the particular error.
  */
 export interface ErrorResponse {
   error?: ErrorDetails;
@@ -211,6 +303,26 @@ export interface Operation {
  */
 export interface UsageDetailsListResult extends Array<UsageDetail> {
   readonly nextLink?: string;
+}
+
+/**
+ * @class
+ * Initializes a new instance of the ReservationSummariesListResult class.
+ * @constructor
+ * Result of listing reservation summaries.
+ *
+ */
+export interface ReservationSummariesListResult extends Array<ReservationSummaries> {
+}
+
+/**
+ * @class
+ * Initializes a new instance of the ReservationDetailsListResult class.
+ * @constructor
+ * Result of listing reservation details.
+ *
+ */
+export interface ReservationDetailsListResult extends Array<ReservationDetails> {
 }
 
 /**